phase6_backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pydantic==2.7.1
pandas==2.2.2
pyarrow==15.0.2
python-dotenv==1.0.1
